Refactor applicant_profile_presenter specs to be FIRST
The tests in this file violated four of the five FIRST principles (because I don't know if they were written before the code). They were not fast as they created a persistent object in the database per test. They also weren't isolated because they had a ton of extraneous attributes that were not related to the methods that were under test. This made it very difficult to understand what was going on without looking at the source code. They were repeatable but not really extensible nor could someone add a test with confidence that running in any order would not break existing tests. There were so many creates that it was impossible to be confident in this test file. Now it is fixed.